1.id값 타입 수정 2.Member 관리기능 추가 3.Member관련 서비스 추가 4.findById 는 orElseThrow로 수정 5.부서원관리 기능 테스트케이스 작성

+++ daou/src/main/java/com/organization/demo/repository/MemberRepository.java
@@ -1,6 +1,7 @@
 package com.organization.demo.repository;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -8,7 +9,9 @@ import org.springframework.stereotype.Repository;
 import com.organization.demo.entity.MemberEntity;
 
 @Repository
-public interface MemberRepository extends JpaRepository<MemberEntity, Integer>{
+public interface MemberRepository extends JpaRepository<MemberEntity, Long>{
 
 	List<MemberEntity> findByNameContains(String name);
+
+//	Optional<MemberEntity> findById(Long id);
 }
